chor(search): split dataset-stats into its own module from report-stats page
